hw/hw4/ai-learning-assistant/requirements.txt
streamlit>=1.30.0
streamlit-agraph>=0.0.45
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
